changed order of modules, plus fixed some typos

+++ code/jee/ejb/framework/proxy/src/main/java/org/tsdes/jee/ejb/framework/proxy/ObjectFactory.java
@@ -14,7 +14,7 @@ public class ObjectFactory {
     }
 
     public int getTotalInvocationCount(){
-        return (int) callLogs.stream()
+        return  callLogs.stream()
                 .mapToInt(CallLog::getInvocationCount)
                 .sum();
     }
